refactor: moved menu to App and fixed menu navigation for tabs
