+++ src/App.svelte
@@ -12,15 +12,16 @@
 	import Toast from './components/Toast.svelte';
 	import Logo from './components/Logo.svelte';
 	import ToggleTheme from './components/ToggleTheme.svelte';
+	import Menu from './components/Menu.svelte';
+
 	import About from './routes/About.svelte';
 	import NotFound from './routes/NotFound.svelte';
 	import Login from './routes/Login.svelte';
 	import Home from './routes/Home.svelte';
 	import { checkAuth, initAuth0Client } from './auth';
-	import { setAuth0Client, getAuth0Client } from './lib/auth-client';
-	import { toastActions, toastMessage, status, isLoggedin } from './stores';
+	import { setAuth0Client } from './lib/auth-client';
+	import { toastActions, toastMessage, status, isLoggedin, currentTab, tabs } from './stores';
 	import themeStore from './stores/theme';
-	import type { ComponentType } from 'svelte';
 
 	let auth0 = $state<Auth0Client>();
 	let wrapper: HTMLElement;
@@ -59,6 +60,17 @@
 		handleBackToTop();
 		if (synced && !auth0) initializeAuth();
 	});
+
+	let menuItems = $derived(
+		tabs.map((item) => ({
+			...item,
+			selected: item.label === $currentTab
+		}))
+	);
+
+	const handleMenu = (path: string) => {
+		currentTab.set(path);
+	};
 
 	const routes = {
 		'/about': About,
@@ -110,6 +122,11 @@
 
 		<ToggleTheme />
 	</header>
+	<aside class="Menu">
+		<Menu {menuItems} goTo={handleMenu}>
+			<Button size="large" variant="link" onclick={() => push('/about')}>About</Button>
+		</Menu>
+	</aside>
 	<Router {routes} />
 	{#if showBackToTop}
 		<footer class="Footer" transition:fly={{ y: 100, duration: 500 }}>
@@ -195,5 +212,12 @@
 		&[data-syncing='ERROR'] {
 			--indicator-color: var(--red);
 		}
+	}
+
+	.Menu {
+		position: fixed;
+		top: 1rem;
+		left: 1rem;
+		z-index: 9;
 	}
 </style>
